# Further reduce Labes() calls in the metrics registry