**What this PR does**:
Here we continue pre-allocating the labels for the registry to reduce memory allocations. This is a follow-on to https://github.com/grafana/tempo/pull/4232, https://github.com/grafana/tempo/pull/4244 and https://github.com/grafana/tempo/pull/4261.

**Which issue(s) this PR fixes**:
Fixes #<issue number>

**Checklist**
- [x] Tests updated
- [ ] Documentation added
- [ ] `CHANGELOG.md` updated - the order of entries should be `[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`