@zalegrala 2024-11-04
```
❯ benchstat main.txt wip.txt
goos: linux
goarch: amd64
pkg: github.com/grafana/tempo/modules/generator
cpu: AMD Ryzen 7 PRO 7840U w/ Radeon 780M Graphics
           │   main.txt   │               wip.txt                │
           │    sec/op    │    sec/op     vs base                │
Collect-16   5.032µ ± 11%   4.117µ ± 22%  -18.18% (p=0.004 n=10)

           │  main.txt   │            wip.txt            │
           │ heap_in_use │ heap_in_use  vs base          │
Collect-16   17.41M ± 1%   17.37M ± 1%  ~ (p=0.669 n=10)

           │   main.txt   │               wip.txt               │
           │     B/op     │     B/op      vs base               │
Collect-16   12.63Ki ± 6%   12.33Ki ± 6%  -2.30% (p=0.050 n=10)

           │   main.txt   │            wip.txt             │
           │  allocs/op   │  allocs/op    vs base          │
Collect-16   17.00 ± 135%   16.00 ± 125%  ~ (p=0.150 n=10)
``